Set uid/gid in /etc/rc.conf on FreeBSD
The FreeBSD rc script for opendkim does privileges dropping before executing
the daemon (which is then not able to change uid/gid).  Manage the
corresponding entries in /etc/rc.conf so that they align with the "UserID"
parameter from opendkim.conf.

This is similar to what the sysconfigfile configuration does on Debian /
RedHat.